requirements.txt
transformers
sentence-transformers
faiss-cpu
datasets
langchain
pandas
